feat: package allow skip app_id

+++ repository/repomodel/user-usable-add-event.go
@@ -1,6 +1,6 @@
 package repomodel
 
 type AvailableSkipAds struct {
-	EventAddID uint32
+	EventAddID uint64
 	Remaining  uint32
 }
